[Refactor] `EditorCommand` is executed based on the context of the current editor
